II Dev Tools: Publishing; Windows Installer - Implemented project "Publishing" to assist in batch functions   - More functional/responsive than .bat file - Targeted Windows Installer to II Avalonia project output

+++ IISE Windows/Controls/PropertyFloat.xaml.cs
@@ -48,7 +48,7 @@ namespace II.Scenario_Editor.Controls {
             numValue.LostFocus += sendPropertyChange;
         }
 
-        public void Set (float value) {
+        public void Set (double value) {
             numValue.ValueChanged -= sendPropertyChange;
             numValue.Value = (decimal)value;
             numValue.ValueChanged += sendPropertyChange;
